requirements.txt
torch
timm
transformers
gradio
numpy
pillow